+++ pypsa/components/_types/buses.py
@@ -63,6 +63,8 @@ class Buses(Components):
     def add_missing_buses(self, **kwargs: Any) -> pd.Index:
         """Add buses that are referenced by components but not yet defined.
 
+        <!-- md:badge-version v1.1.0 -->
+
         Parameters
         ----------
         **kwargs : Any
@@ -80,7 +82,7 @@ class Buses(Components):
         >>> n.components.buses.add_missing_buses(v_nom=100.0)
         Index(['my_bus'], dtype='object')
 
-        Buses are added without the need to extra call `n.add`:
+        Buses are added without needing to call `n.add` separately:
 
         >>> n.components.buses.static  # doctest: +ELLIPSIS
                 v_nom type    x    y  ...
@@ -94,7 +96,6 @@ class Buses(Components):
         for c in self.n_save.c.values():
             if c.static.empty:
                 continue
-            # Get bus columns for this component
             bus_cols = c.static.columns[c.static.columns.str.contains(RE_PORTS_FILTER)]
             for attr in bus_cols:
                 buses = c.static[attr].astype(str)
@@ -104,7 +105,7 @@ class Buses(Components):
                 # Filter empty strings for global constraints
                 if c.name == "GlobalConstraint":
                     buses = buses[buses != ""]
-                # Find missing buses
+
                 missing = ~buses.isin(self.n_save.c.buses.names)
                 all_buses.update(buses[missing & (buses != "") & (buses != "nan")])